implemented smart contract event polling and processing

+++ Backend/Application/Events/ValueSetEvent/ValueSetEvent.cs
@@ -1,4 +1,3 @@
-using System.Numerics;
 using Nethereum.ABI.FunctionEncoding.Attributes;
 
 namespace Application.Events;
@@ -6,9 +5,9 @@ namespace Application.Events;
 [Event("ValueSet")]
 public class ValueSetEvent : IEventDTO
 {
-    [Parameter("uint", "oldValue", 1, true)]
+    [Parameter("uint256", "oldValue", 1, true)]
     public uint OldValue { get; set; }
 
-    [Parameter("uint", "newValue", 2, true)]
+    [Parameter("uint256", "newValue", 2, true)]
     public uint NewValue { get; set; }
 }
